pyproject.toml
[tool.poetry]
name = "004_API_Devman"
version = "0.1.0"
description = ""
authors = ["Roman Okatev <nebulawalker@live.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
